test: 속성 기반 테스트 추가
